@@ -1,7 +1,7 @@
 {
   "name": "email-securely-app",
   "description": "Unofficial desktop app for E2E encrypted email providers",
-  "version": "2.2.1",
+  "version": "2.3.0",
   "author": "Vladimir Yakovlev <desktop-app@protonmail.ch>",
   "license": "MIT",
   "homepage": "https://github.com/vladimiry/email-securely-app",
@@ -108,7 +108,7 @@
     "html-to-text": "4.0.0",
     "image-processing-js": "0.0.3",
     "js-base64": "2.5.1",
-    "keytar": "4.3.0",
+    "keytar": "4.3.2",
     "memory-streams": "0.1.3",
     "mime-types": "2.1.21",
     "msgpack-lite": "0.1.26",
@@ -126,27 +126,27 @@
     "valid-url": "1.0.9"
   },
   "devDependencies": {
-    "@angular-devkit/build-optimizer": "0.12.3",
-    "@angular/animations": "7.2.2",
-    "@angular/common": "7.2.2",
-    "@angular/compiler": "7.2.2",
-    "@angular/compiler-cli": "7.2.2",
-    "@angular/core": "7.2.2",
-    "@angular/forms": "7.2.2",
-    "@angular/http": "7.2.2",
-    "@angular/language-service": "7.2.2",
-    "@angular/platform-browser": "7.2.2",
-    "@angular/platform-browser-dynamic": "7.2.2",
-    "@angular/router": "7.2.2",
+    "@angular-devkit/build-optimizer": "0.13.0",
+    "@angular/animations": "7.2.3",
+    "@angular/common": "7.2.3",
+    "@angular/compiler": "7.2.3",
+    "@angular/compiler-cli": "7.2.3",
+    "@angular/core": "7.2.3",
+    "@angular/forms": "7.2.3",
+    "@angular/http": "7.2.3",
+    "@angular/language-service": "7.2.3",
+    "@angular/platform-browser": "7.2.3",
+    "@angular/platform-browser-dynamic": "7.2.3",
+    "@angular/router": "7.2.3",
     "@angularclass/hmr": "2.1.3",
     "@email-securely-app/import-sort-style": "0.1.0",
     "@expo/spawn-async": "1.4.0",
-    "@ng-select/ng-select": "2.15.1",
+    "@ng-select/ng-select": "2.15.2",
     "@ngrx/effects": "7.2.0",
     "@ngrx/router-store": "7.2.0",
     "@ngrx/store": "7.2.0",
-    "@ngtools/webpack": "7.2.3",
-    "@octokit/rest": "16.13.4",
+    "@ngtools/webpack": "7.3.0",
+    "@octokit/rest": "16.14.1",
     "@types/angular": "1.6.53",
     "@types/byline": "4.2.31",
     "@types/chalk": "2.2.0",
@@ -159,9 +159,10 @@
     "@types/keytar": "4.0.1",
     "@types/mime-types": "2.1.0",
     "@types/mini-css-extract-plugin": "0.2.0",
+    "@types/mithril-global": "1.1.1",
     "@types/mkdirp": "0.5.2",
     "@types/msgpack-lite": "0.1.6",
-    "@types/node": "10.12.18",
+    "@types/node": "10.12.21",
     "@types/node-fetch": "2.1.4",
     "@types/p-queue": "3.0.0",
     "@types/path-is-inside": "1.0.0",
@@ -182,7 +183,7 @@
     "@types/webdriverio": "4.13.3",
     "@types/webpack": "4.4.24",
     "@types/webpack-dev-server": "3.1.1",
-    "@types/webpack-env": "1.13.6",
+    "@types/webpack-env": "1.13.7",
     "@types/webpack-merge": "4.1.3",
     "@types/webpack-node-externals": "1.6.3",
     "@vladimiry/unionize": "2.1.2-add-tagprefix-option",
@@ -202,7 +203,7 @@
     "cssnano": "4.1.8",
     "devtron": "1.4.0",
     "dugite": "1.82.0",
-    "electron": "4.0.3",
+    "electron": "4.0.4",
     "electron-builder": "20.38.5",
     "exports-loader": "0.7.0",
     "fast-glob": "2.2.6",
@@ -224,7 +225,7 @@
     "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "4.0.0-rc.2",
     "less-loader": "4.1.0",
-    "lint-staged": "8.1.1",
+    "lint-staged": "8.1.3",
     "mini-css-extract-plugin": "0.5.0",
     "ndx": "1.0.0-beta.2",
     "ndx-query": "1.0.0-beta.2",
@@ -264,7 +265,7 @@
     "tslint-consistent-codestyle": "1.15.0",
     "tslint-eslint-rules": "5.4.0",
     "tslint-rules-bunch": "0.0.7",
-    "typescript": "3.2.4",
+    "typescript": "3.3.1",
     "uglifyjs-webpack-plugin": "2.1.1",
     "url-loader": "1.1.2",
     "webpack": "4.29.0",
